Get rid of the separation between __init__.py and vector.py
